Refactor endpoint email/sendmail
