+++ package.json
@@ -23,6 +23,7 @@
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-fileupload": "^1.1.6",
     "hbs": "^4.0.6",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.8.9",
